requirements.txt
asgiref==3.4.1
Django==3.0.1
djangorestframework==3.11.0
djongo==1.3.6
dnspython==2.1.0
jsonfield==3.1.0
mongoengine==0.23.1
psycopg2-binary==2.8.6
pymongo==3.12.3
pytz==2021.3
sqlparse==0.2.4